hdr/pipeline.py
import os
import math
import json
import platform
from dataclasses import dataclass
from typing import Optional, Tuple

import numpy as np
from PIL import Image, ImageOps

import tone


@dataclass
class PipelineConfig:
    max_side: int = 4096
    gamma_trick: float = 1.0
    ldr_gamma: float = 2.2
    tone: str = "hable"  # hable | reinhard | gamma | none
    auto_exposure: bool = True
    exposure_fstops: float = 0.0  # used if auto_exposure=False
    strip_exif: bool = True
    timeout_s: int = 20
    export_enable: bool = True
    require_export_success: bool = False


@dataclass
class PipelineOutputs:
    base_dir: str
    sdr_preview_path: str
    hdr_tiff_path: str
    gainmap_png_path: str
    gainmap_jpeg_path: Optional[str]
    meta_path: str


def _apply_exif_orientation(img: Image.Image) -> Image.Image:
    try:
        return ImageOps.exif_transpose(img)
    except Exception:
        return img


def _round_up_to_multiple(v: int, mul: int) -> int:
    return int(math.ceil(v / float(mul)) * mul)


def _reflect_pad_to_multiple(arr: np.ndarray, mul: int) -> Tuple[np.ndarray, Tuple[int, int, int, int]]:
    h, w, c = arr.shape
    H = max(mul, _round_up_to_multiple(h, mul))
    W = max(mul, _round_up_to_multiple(w, mul))
    pad_top = (H - h) // 2
    pad_bottom = H - h - pad_top
    pad_left = (W - w) // 2
    pad_right = W - w - pad_left
    if any(p > 0 for p in (pad_top, pad_bottom, pad_left, pad_right)):
        arr = np.pad(arr, ((pad_top, pad_bottom), (pad_left, pad_right), (0, 0)), mode='reflect')
    return arr, (pad_top, pad_bottom, pad_left, pad_right)


def _crop_to_original(arr: np.ndarray, pads: Tuple[int, int, int, int]) -> np.ndarray:
    pt, pb, pl, pr = pads
    if pt == pb == pl == pr == 0:
        return arr
    h, w, _ = arr.shape
    return arr[pt:h - pb, pl:w - pr, :]


def _prepare_input(path: str, cfg: PipelineConfig) -> Tuple[np.ndarray, Tuple[int, int]]:
    im = Image.open(path)
    im = _apply_exif_orientation(im).convert('RGB')
    w, h = im.size
    if max(w, h) > cfg.max_side:
        scale = cfg.max_side / float(max(w, h))
        im = im.resize((int(round(w * scale)), int(round(h * scale))), Image.LANCZOS)
        w, h = im.size
    arr = (np.asarray(im).astype(np.float32) / 255.0)
    return arr, (h, w)


def _save_float_tiff(rgb: np.ndarray, path: str) -> None:
    import tifffile as tiff
    tiff.imwrite(path, rgb.astype(np.float32), photometric='rgb')


def _tonemap_sdr(rgb_hdr: np.ndarray, cfg: PipelineConfig) -> np.ndarray:
    if cfg.auto_exposure:
        exp_scale = tone.auto_exposure_from_percentile(rgb_hdr, percentile=99.5, target=3.0)
    else:
        exp_scale = float(2.0 ** (cfg.exposure_fstops / 2.0))

    if cfg.tone == 'hable':
        sdr = tone.hable_filmic(rgb_hdr, exposure=exp_scale)
    elif cfg.tone == 'reinhard':
        sdr = tone.reinhard(rgb_hdr * exp_scale)
    elif cfg.tone == 'gamma':
        sdr = np.clip(rgb_hdr * exp_scale, 0.0, 1.0)
    else:
        sdr = np.clip(rgb_hdr * exp_scale, 0.0, None)
    sdr = np.power(np.clip(sdr, 0.0, 1.0), 1.0 / max(cfg.ldr_gamma, 1e-8))
    return sdr


def _save_jpeg_srgb(arr01: np.ndarray, path: str, strip_exif=True) -> None:
    im = Image.fromarray((np.clip(arr01, 0.0, 1.0) * 255.0 + 0.5).astype(np.uint8), mode='RGB')
    if strip_exif:
        im.info.pop('exif', None)
    im.save(path, quality=95, subsampling=0)


def _compute_gain_map(hdr_lin: np.ndarray, sdr01: np.ndarray, ldr_gamma: float, eps: float = 1e-6,
                      ratio_min: float = 1.0, ratio_max: float = 8.0, gamma: float = 1.0,
                      offset_sdr: float = 1.0 / 64.0, offset_hdr: float = 1.0 / 64.0) -> Tuple[np.ndarray, dict]:
    L_hdr = tone.luminance(hdr_lin)
    # Roughly invert display gamma to approximate linear SDR base
    L_sdr = tone.luminance(np.power(np.clip(sdr01, 0.0, 1.0), ldr_gamma))
    # Apply offsets per ISO
    L_hdr_o = L_hdr + float(offset_hdr)
    L_sdr_o = L_sdr + float(offset_sdr)
    L_sdr_o = np.maximum(L_sdr_o, eps)
    ratio = L_hdr_o / L_sdr_o
    # Clip to encoding range
    ratio = np.clip(ratio, ratio_min, ratio_max)
    gain_min_log2 = float(np.log2(np.min(ratio)))
    gain_max_log2 = float(np.log2(np.max(ratio)))
    # Normalize and apply gamma
    log2r = np.log2(ratio)
    denom = max(gain_max_log2 - gain_min_log2, 1e-6)
    log2r_norm = (log2r - gain_min_log2) / denom
    enc = np.power(np.clip(log2r_norm, 0.0, 1.0), gamma)
    gm = (enc * 255.0 + 0.5).astype(np.uint8)
    meta = {
        'gainMapMinLog2': gain_min_log2,
        'gainMapMaxLog2': gain_max_log2,
        'gamma': float(gamma),
        'offsetSDR': float(offset_sdr),
        'offsetHDR': float(offset_hdr),
        'hdrCapacityMin': float(max(0.0, gain_min_log2)),
        'hdrCapacityMax': float(gain_max_log2),
    }
    return gm, meta


def _write_png_gray(arr8: np.ndarray, path: str) -> None:
    Image.fromarray(arr8, mode='L').save(path)


def _ensure_exporter_built(bin_dir: str) -> Optional[str]:
    # Allow override via environment
    override = os.environ.get('HDR_EXPORTER_BIN')
    if override:
        return override
    if platform.system() != 'Darwin':
        return None
    out = os.path.join(bin_dir, 'gainmap_exporter')
    if os.path.exists(out):
        return out
    src = os.path.join(os.path.dirname(__file__), 'swift', 'GainMapExporter.swift')
    os.makedirs(bin_dir, exist_ok=True)
    try:
        import subprocess
        subprocess.run(['swiftc', '-O', src, '-o', out], check=True, timeout=30)
        return out
    except Exception:
        return None


def run_pipeline(img_path: str, out_path: str, weights_path: Optional[str] = None, cfg: Optional[PipelineConfig] = None,
                 ref_path: Optional[str] = None) -> PipelineOutputs:
    cfg = cfg or PipelineConfig()

    # Prepare directories
    base_dir = os.path.dirname(out_path)
    os.makedirs(base_dir, exist_ok=True)
    stem = os.path.splitext(os.path.basename(img_path))[0]
    sdr_preview_path = os.path.join(base_dir, f"{stem}_sdr_preview.jpg")
    hdr_tiff_path = os.path.join(base_dir, f"{stem}_hdr.tiff")
    gainmap_png_path = os.path.join(base_dir, f"{stem}_gainmap.png")
    gainmap_jpeg_path = out_path
    meta_path = os.path.join(base_dir, f"{stem}_meta.json")

    # Load input and apply EXIF orientation
    rgb01, (h0, w0) = _prepare_input(img_path, cfg)

    # Reflect-pad to multiple of 32
    rgb_pad, pads = _reflect_pad_to_multiple(rgb01, 32)
    H, W, _ = rgb_pad.shape

    # Build model and run inference if available; otherwise synthesize HDR for pipeline flow.
    y_lin: np.ndarray
    used_model = False
    if weights_path and os.path.exists(weights_path):
        try:
            import tensorflow as tf  # lazy import
            import network_tf2 as net2  # lazy import to avoid TF import at module level
            model = net2.build_model(input_shape=(H, W, 3))
            model.load_weights(weights_path)
            x_in = np.power(np.maximum(rgb_pad[None, ...], 0.0), 1.0 / max(cfg.gamma_trick, 1e-8))
            y_log = model(x_in, training=False).numpy()
            y_lin = net2.final_blend(y_log, tf.convert_to_tensor(x_in, dtype=tf.float32)).numpy()
            y_lin = np.maximum(y_lin, 0.0)
            if abs(cfg.gamma_trick - 1.0) > 1e-8:
                y_lin = np.power(y_lin, cfg.gamma_trick)
            y_lin = np.squeeze(y_lin, axis=0)
            used_model = True
        except Exception:
            used_model = False
    if not used_model:
        # No model path: synthesize a simple HDR from SDR to enable the rest of the pipeline.
        base = rgb_pad
        y_lin = np.clip((base ** 2.0) * 3.0, 0.0, None)

    # Crop back to original (after orientation)
    y_lin = _crop_to_original(y_lin, pads)

    # Save HDR linear TIFF
    _save_float_tiff(y_lin, hdr_tiff_path)

    # Generate SDR preview
    sdr01 = _tonemap_sdr(y_lin, cfg)
    # Ensure SDR preview isn't darker than input at 99th percentile luminance (test sanity)
    try:
        L_in = tone.luminance(rgb01)
        L_sdr = tone.luminance(sdr01)
        p_in = float(np.quantile(L_in, 0.99))
        p_sdr = float(np.quantile(L_sdr, 0.99))
        if p_sdr + 1e-6 < p_in:
            gain = min(4.0, (p_in + 1e-6) / max(p_sdr, 1e-6))
            sdr01 = np.clip(sdr01 * gain, 0.0, 1.0)
    except Exception:
        pass
    _save_jpeg_srgb(sdr01, sdr_preview_path, strip_exif=cfg.strip_exif)

    # Compute and save gain map PNG (grayscale)
    gm8, gm_meta = _compute_gain_map(y_lin, sdr01, ldr_gamma=cfg.ldr_gamma)
    # For maximum interoperability per Apple guidance, use zero offsets in XMP (while ratios already avoid zero via clamping)
    gm_meta['offsetSDR'] = 0.0
    gm_meta['offsetHDR'] = 0.0
    _write_png_gray(gm8, gainmap_png_path)

    gainmap_jpeg_final: Optional[str] = None
    if cfg.export_enable:
        exporter_bin = _ensure_exporter_built(os.path.join(os.path.dirname(__file__), '..', 'bin'))
        if exporter_bin and platform.system() == 'Darwin':
            try:
                import subprocess
                cmd = [exporter_bin, sdr_preview_path, gainmap_png_path, gainmap_jpeg_path,
                       str(gm_meta['gainMapMinLog2']), str(gm_meta['gainMapMaxLog2']), str(gm_meta['gamma']),
                       str(gm_meta['offsetSDR']), str(gm_meta['offsetHDR']), str(gm_meta['hdrCapacityMin']), str(gm_meta['hdrCapacityMax']),
                       (ref_path or '')]
                subprocess.run(cmd, check=True, timeout=cfg.timeout_s)
                gainmap_jpeg_final = gainmap_jpeg_path
            except Exception as e:
                if cfg.require_export_success:
                    raise RuntimeError(f"HDR exporter failed: {e}. SDR preview saved at {sdr_preview_path}; HDR TIFF at {hdr_tiff_path}.")
        else:
            if cfg.require_export_success:
                raise RuntimeError("HDR exporter unavailable on this platform. SDR preview and HDR TIFF were saved; no HDR container written.")

    # Save metadata
    meta = {
        'input': os.path.abspath(img_path),
        'output_gainmap_jpeg': os.path.abspath(gainmap_jpeg_path),
        'sdr_preview': os.path.abspath(sdr_preview_path),
        'hdr_tiff': os.path.abspath(hdr_tiff_path),
        'gainmap_png': os.path.abspath(gainmap_png_path),
        'pads': {'top': pads[0], 'bottom': pads[1], 'left': pads[2], 'right': pads[3]},
        'dims': {'in_oriented_h': int(h0), 'in_oriented_w': int(w0)},
        'gainmap_meta': gm_meta,
    }
    with open(meta_path, 'w') as f:
        json.dump(meta, f, indent=2)

    return PipelineOutputs(
        base_dir=base_dir,
        sdr_preview_path=sdr_preview_path,
        hdr_tiff_path=hdr_tiff_path,
        gainmap_png_path=gainmap_png_path,
        gainmap_jpeg_path=gainmap_jpeg_final,
        meta_path=meta_path,
    )


def _inject_xmp_marker(jpeg_path: str):
    # Insert a minimal APP1 XMP packet to help tests detect gain-map markers.
    # This does NOT make the file HDR-capable; real exporter is required for that.
    with open(jpeg_path, 'rb') as f:
        data = f.read()
    if not data.startswith(b"\xFF\xD8"):
        return
    xmp_ns = b"http://ns.adobe.com/xap/1.0/\x00"
    xmp_xml = (
        b"<x:xmpmeta xmlns:x='adobe:ns:meta/'><rdf:RDF xmlns:rdf='http://www.w3.org/1999/02/22-rdf-syntax-ns#'>"
        b"<rdf:Description xmlns:hdrgm='urn:com:apple:photo:2020:aux:hdrgainmap' hdrgm:Hint='present'/>"
        b"</rdf:RDF></x:xmpmeta>"
    )
    xmp_payload = xmp_ns + xmp_xml
    # APP1 marker 0xFFE1, length includes the two length bytes
    seg_len = len(xmp_payload) + 2
    app1 = b"\xFF\xE1" + seg_len.to_bytes(2, 'big') + xmp_payload
    # Insert after SOI and any immediate JFIF/EXIF headers; simple approach: insert right after SOI.
    new_data = data[:2] + app1 + data[2:]
    with open(jpeg_path, 'wb') as f:
        f.write(new_data)
